新增订单创建功能，包含订单请求类和相关服务逻辑

+++ src/main/java/org/cjj/saaassistant/pojo/Order.java
@@ -3,7 +3,7 @@ package org.cjj.saaassistant.pojo;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -13,7 +13,7 @@ import java.util.List;
 public class Order {
     private Integer id;
     private Integer userId; // 下单用户id
-    private Integer totalPrice; // 订单总价
+    private BigDecimal totalPrice; // 订单总价
     private String status; // 支付状态
     private Date createTime; // 创建时间
     private Date updateTime; // 更新时间
