+++ src/main/java/org/cjj/saaassistant/pojo/OrderProduct.java
@@ -3,7 +3,7 @@ package org.cjj.saaassistant.pojo;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -13,8 +13,8 @@ public class OrderProduct {
     private Integer id; // 订单项id
     private Integer orderId; // 关联订单
     private Integer productId; // 关联商品
-    private String productName;
-    private Integer productPrice;
+    private String productName; // 商品名称快照
+    private BigDecimal productPrice; // 商品价格快照
     private Integer quantity; // 购买数量
     private Date createTime; // 创建时间
     private Date updateTime; // 更新时间
